requirements.txt
PyCrypto
NumPy
construct
click